Clicking on an element of the mole list now brings up that mole's information.

+++ src/cs310w10/MoleFinder/Controller/MoleController.java
@@ -20,7 +20,7 @@ public class MoleController {
 		return mole;
 	}
 	
-	public static Mole getMoleFromId(int id){
+	public static Mole getMoleFromId(long id){
 	        MolesDataSource source = new MolesDataSource(MoleFinderApplication.getAppContext());
 	        source.open();
 	        Mole mole = source.getMoleFromId(id);
